update ply test
