--- include/icp/parse_ply.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <vector>
-#include <string>
-#include "icp/geo.h"
-
-std::vector<icp::Vector> parse_ply(const std::string& path);
